% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_matching.r
\name{describe}
\alias{describe}
\title{Describe HILDA variables}
\usage{
describe(variables)
}
\arguments{
\item{variables}{A character vector of variable names}
}
\value{
A data table subset of the data \code{\link{dictionary}}
}
\description{
Matches variable names to data dictionary entries and returns a
data table of descriptions. The function is not case sensitive and
matches with or without wave prefixes.
}
\examples{

describe("XHHRAID")
describe(c("pmhlyr", "_TIFDIF"))
}
