% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda.bayesfactor}
\alias{hilda.bayesfactor}
\title{Apply HiLDA to statistically testing the global difference in burdens of mutation signatures between two groups}
\usage{
hilda.bayesfactor(inputG = G, inputParam = Param, refGroup,
  sigOrder = NULL, n.iter = 2000, n.burnin = 0, prob_M1 = 0.5, ...)
}
\arguments{
\item{inputG}{a MutationFeatureData S4 class output by the pmsignature.}

\item{inputParam}{a EstimatedParameters S4 class output by the pmsignature.}

\item{refGroup}{the indice indicating the samples in the reference group.}

\item{sigOrder}{the order of the mutational signatures.}

\item{n.iter}{number of total iterations per chain (default: 2000).}

\item{n.burnin}{length of burn (default: 0).}

\item{prob_M1}{the probability of sampling the model 1 (no difference in the means)}
}
\description{
Apply HiLDA to statistically testing the global difference in burdens of mutation signatures between two groups
}
