% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda.inits}
\alias{hilda.inits}
\title{Estimate initial values for HiLDA test from the pmsignature result}
\usage{
hilda.inits(inputParam = Param, refGroup, sigOrder = NULL, ...)
}
\arguments{
\item{refGroup}{the indice indicating the samples in the reference group.}

\item{sigOrder}{the reference signature should be replaced on the first.}

\item{input}{a EstimatedParameters S4 class output by the pmsignature..}

\item{numSig}{the number of mutation signatures.}
}
\description{
Estimate initial values for HiLDA test from the pmsignature result
}
