% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda.plotSignature}
\alias{hilda.plotSignature}
\title{Read the raw mutation data with the mutation feature vector format, estimate and plot both mutation signatures and their fractions}
\usage{
hilda.plotSignature(inputParam)
}
\arguments{
\item{numBases}{the number of flanking bases around the mutated position.}

\item{numSig}{the number of mutation signatures.}
}
\description{
Read the raw mutation data with the mutation feature vector format, estimate and plot both mutation signatures and their fractions
}
