% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda.posterior}
\alias{hilda.posterior}
\title{Extract the posterior distributions of the mean differences in muational exposures}
\usage{
hilda.posterior(jags.output)
}
\arguments{
\item{jags.output}{the output jags file generated by the jags function from the R2jags package.}
}
\description{
Extract the posterior distributions of the mean differences in muational exposures
}
