% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda.test}
\alias{hilda.test}
\title{Apply HiLDA to statistically testing the burdens of mutation signatures between two groups}
\usage{
hilda.test(inputG = G, inputParam = Param, refGroup, sigOrder = NULL,
  n.iter = 2000, n.burnin = 0, ...)
}
\arguments{
\item{refGroup}{the indice indicating the samples in the reference group.}

\item{input}{a MutationFeatureData S4 class output by the pmsignature.}

\item{numBases}{the number of flanking bases around the mutated position.}

\item{numSig}{the number of mutation signatures.}
}
\description{
Apply HiLDA to statistically testing the burdens of mutation signatures between two groups
}
