% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda_bayesfactor_result}
\alias{hilda_bayesfactor_result}
\title{Compute the Bayes factor}
\usage{
hilda_bayesfactor_result(jags.output, prob_M1 = 0.5)
}
\arguments{
\item{jags.output}{the output jags file generated by the jags function from the R2jags package.}

\item{prob_M1}{the probability of sampling the model 1 (no difference in the means), default is 0.5}
}
\description{
Compute the Bayes factor
}
