% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda_rhat}
\alias{hilda_rhat}
\title{Output the maximum potential scale reduction statistic of all parameters estimated}
\usage{
hilda_rhat(jags.output)
}
\arguments{
\item{jags.output}{the output jags file generated by the jags function from the R2jags package.}
}
\description{
Output the maximum potential scale reduction statistic of all parameters estimated
}
