% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_functions.R
\name{hilda_test}
\alias{hilda_test}
\title{Apply HiLDA to statistically testing the burdens of mutation signatures between two groups}
\usage{
hilda_test(inputG = G, inputParam = Param, refGroup, sigOrder = NULL,
  n.iter = 2000, n.burnin = 0, ...)
}
\arguments{
\item{inputG}{a MutationFeatureData S4 class output by the pmsignature.}

\item{inputParam}{a EstimatedParameters S4 class output by the pmsignature.}

\item{refGroup}{the reference group}

\item{sigOrder}{the indice indicating the samples in the reference group.}

\item{n.iter}{number of total iterations per chain (default: 2000).}

\item{n.burnin}{length of burn (default: 0).}
}
\description{
Apply HiLDA to statistically testing the burdens of mutation signatures between two groups
}
