% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_diagnosis.R
\name{hildaLocalResult}
\alias{hildaLocalResult}
\title{Extract the posterior distributions of the mean differences in muational
exposures}
\usage{
hildaLocalResult(jagsOutput)
}
\arguments{
\item{jagsOutput}{the output jags file generated by the jags function from
the R2jags package.}
}
\value{
a data frame that contains the posterior distributions of difference.
}
\description{
Extract the posterior distributions of the mean differences in muational
exposures
}
\examples{

load(system.file("extdata/hildaLocal.rdata", package="HiLDA"))
hildaLocalResult(hildaLocal)

}
