% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilda_local.R
\name{hildaLocalTest}
\alias{hildaLocalTest}
\title{Apply HiLDA to statistically testing the burdens of mutation signatures
between two groups}
\usage{
hildaLocalTest(inputG, numSig, refGroup, useInits = NULL,
  sigOrder = NULL, nIter = 2000, nBurnin = 0, ...)
}
\arguments{
\item{inputG}{a MutationFeatureData S4 class output by the pmsignature.}

\item{numSig}{an integer number of the number of mutational signatures.}

\item{refGroup}{the reference group.}

\item{useInits}{a EstimatedParameters S4 class output by the pmsignature
(default: NULL)}

\item{sigOrder}{the indice indicating the samples in the reference group
(default: NULL)}

\item{nIter}{number of total iterations per chain (default: 2000).}

\item{nBurnin}{length of burn (default: 0).}

\item{...}{Other arguments passed on to methods.}
}
\value{
the output jags file
}
\description{
Apply HiLDA to statistically testing the burdens of mutation signatures
between two groups
}
\examples{

load(system.file("extdata/sample.rdata", package="HiLDA"))

## with initial values
hildaLocal <- hildaLocalTest(inputG=G, numSig=3, refGroup=1:4, nIter=1000)

}
