% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_plot.R
\name{pmPlotSignature}
\alias{pmPlotSignature}
\title{Plot mutation signatures from pmsignature output}
\usage{
pmPlotSignature(inputParam, sigOrder = NULL, colorList = NULL, ...)
}
\arguments{
\item{inputParam}{a estimatedParameters S4 class output by the pmsignature.}

\item{sigOrder}{the order of signatures if needed (default: NULL).}

\item{colorList}{a list of color to highlight the signatures (default: NULL).}

\item{...}{Other arguments passed on to methods.}
}
\value{
a plot object containing all mutational signatures
}
\description{
Plot mutation signatures from pmsignature output
}
\examples{

load(system.file("extdata/sampleParam.rdata", package="HiLDA"))

pmPlotSignature(Param)

}
