% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read.hilda}
\alias{read.hilda}
\title{Read & combine HILDA files}
\usage{
read.hilda(paths, variables)
}
\arguments{
\item{paths}{a character vector of file paths to HILDA files}

\item{variables}{a character vector variables names to parse. These can use}
}
\value{
data.table
}
\description{
This function reads in any HILDA file (SAS, Stata and SPSS included) and merges
the files where appropriate. This function works best with HILDA's TAB files as
they allow for parsing of individual columns, not the entire dataset. This function
is designed for HILDA 16 general and restricted and may not function correctly with
other versions of HILDA.
}
\examples{

}
