% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_getSignature.R
\name{updatePMSParam}
\alias{updatePMSParam}
\title{A function for updating parameters using EM-algorithm}
\usage{
updatePMSParam(p, y)
}
\arguments{
\item{p}{this variable includes the parameters for mutation signatures and 
membership parameters}

\item{y}{this variable includes the information on the mutation features, 
the number of mutation signatures specified and so on}
}
\value{
a value
}
\description{
A function for updating parameters using EM-algorithm
}
